+++ apps/wallet/src/ui/app/components/ledger/ConnectLedgerModalContainer.tsx
@@ -14,6 +14,8 @@ export function ConnectLedgerModalContainer() {
 	const accountsUrl = useNextMenuUrl(true, '/accounts');
 	const importLedgerAccountsUrl = useNextMenuUrl(true, '/import-ledger-accounts');
 
+	console.log(accountsUrl, importLedgerAccountsUrl);
+
 	return (
 		<ConnectLedgerModal
 			onClose={() => {
