+++ apps/wallet/src/ui/app/index.tsx
@@ -1,10 +1,11 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useFeatureIsOn } from '@growthbook/growthbook-react';
+// import { useFeatureIsOn } from '@growthbook/growthbook-react';
 import { useEffect } from 'react';
 import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 
+import { ImportLedgerAccounts } from './components/ledger/ImportLedgerAccounts';
 import { useInitialPageView } from './hooks/useInitialPageView';
 import AssetsPage from './pages/home/assets';
 import { QredoConnectInfoPage } from './pages/qredo-connect/QredoConnectInfoPage';
@@ -15,8 +16,12 @@ import { Staking } from './staking/home';
 import ForgotPasswordPage from '_app/wallet/forgot-password-page';
 import LockedPage from '_app/wallet/locked-page';
 import { useAppDispatch, useAppSelector } from '_hooks';
+import { AccountsPage } from '_pages/accounts/AccountsPage';
+import { AddAccountPage } from '_pages/accounts/AddAccountPage';
+import { ImportPassphrasePage } from '_pages/accounts/ImportPassphrasePage';
+import { ImportPrivateKeyPage } from '_pages/accounts/ImportPrivateKeyPage';
 import { ApprovalRequestPage } from '_pages/approval-request';
-import WelcomePageV2 from '_pages/enoki-onboarding/WelcomePage';
+import { WelcomePage as WelcomePageV2 } from '_pages/enoki-onboarding/WelcomePage';
 import HomePage, {
 	TokensPage,
 	TransactionBlocksPage,
@@ -48,7 +53,7 @@ const HIDDEN_MENU_PATHS = [
 
 const App = () => {
 	const dispatch = useAppDispatch();
-	const useNewOnboardingFlow = useFeatureIsOn('enoki-social-sign-in');
+	const useNewOnboardingFlow = true; // ;useFeatureIsOn('enoki-social-sign-in');
 	const isPopup = useAppSelector((state) => state.app.appType === AppType.popup);
 	useEffect(() => {
 		document.body.classList.remove('app-initializing');
@@ -86,6 +91,15 @@ const App = () => {
 			</Route>
 
 			<Route path="welcome" element={useNewOnboardingFlow ? <WelcomePageV2 /> : <WelcomePage />} />
+			{useNewOnboardingFlow && (
+				<Route path="/accounts" element={<AccountsPage />}>
+					<Route path="add-account" element={<AddAccountPage />} />
+					<Route path="create-new-account" element={<AddAccountPage />} />
+					<Route path="import-ledger-accounts" element={<ImportLedgerAccounts />} />
+					<Route path="import-passphrase" element={<ImportPassphrasePage />} />
+					<Route path="import-private-key" element={<ImportPrivateKeyPage />} />
+				</Route>
+			)}
 			<Route path="/initialize" element={<InitializePage />}>
 				<Route path="select" element={<SelectPage />} />
 				<Route path="create" element={<CreatePage />} />
