+++ apps/wallet/src/ui/app/pages/accounts/ImportPrivateKeyPage.tsx
@@ -1,0 +1,6 @@
+// Copyright (c) Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+export function ImportPrivateKeyPage() {
+	return <div>Implement me!</div>;
+}
